rename some files and vars
