+++ packages/core/http/core-http-server-internal/src/static_assets.ts
@@ -8,7 +8,7 @@
 
 import { Env } from '@kbn/config';
 import type { BasePath } from './base_path_service';
-import { CdnConfig } from './cdn';
+import { CdnConfig } from './cdn_config';
 
 export interface InternalStaticAssets {
   getHrefBase(): string;
